feat(usersProviders): order wrappers by platform

+++ src/data_provider/src/extended_sources/users/tests/test_logged_in_users_unix.cpp
@@ -7,9 +7,11 @@
  * Foundation.
  */
 
-#include "logged_in_users_unix.hpp"
 #include "gtest/gtest.h"
 #include "gmock/gmock.h"
+
+#include "logged_in_users_unix.hpp"
+#include "iutmpx_wrapper.hpp"
 
 class MockUtmpxWrapper : public IUtmpxWrapper
 {
@@ -42,7 +44,7 @@ TEST(LoggedInUsersProviderTest, CollectReturnsExpectedJson)
     LoggedInUsersProvider provider(mockWrapper);
     auto result = provider.collect();
 
-    ASSERT_EQ(result.size(), 1);
+    ASSERT_EQ(result.size(), static_cast<size_t>(1));
     EXPECT_EQ(result[0]["user"], "testuser");
     EXPECT_EQ(result[0]["type"], "user");
     EXPECT_EQ(result[0]["tty"], "pts/0");
